# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级在线考试系统，支持考题管理、考试配置、考生管理、成绩统计与导出。\n认证方式：Bearer Token（通过 /api/auth/login 获取，Cookie 或 Authorization Header 传递）",
        "title": "在线考试系统 API",
        "termsOfService": "http://svip.com/terms",
        "contact": {
            "name": "API 支持",
            "email": "admin@svip.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/scores": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "分页查询所有考生的考试成绩，支持按关键词、分类筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "获取成绩列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词（考生姓名/考试名称）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类 ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码 (默认 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数 (默认 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成绩列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PagedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/scores/export": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "按分类导出成绩为 CSV 或 Excel 文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
                    "text/csv"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "导出成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类 ID（为空则导出所有）",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "导出格式：csv 或 xlsx（默认 xlsx）",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成绩文件下载",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "导出失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/exam/questions": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "学生端获取考试题目（不含答案）。如果已提交且不允许重复，则返回历史成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生端"
                ],
                "summary": "获取考试题目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考试 ID",
                        "name": "examId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "考生姓名",
                        "name": "studentName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "考题列表或历史成绩",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ExamQuestionsData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "缺少 examId",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "考试不存在",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/exam/submit": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "学生提交考试答案，系统自动判分并返回结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生端"
                ],
                "summary": "提交考试答案",
                "parameters": [
                    {
                        "description": "答案数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubmitExamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "判分结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ExamResultRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或不允许重复提交",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/student/exams": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "学生端查看当前可参加的考试（状态为 active/upcoming），含是否已参加的标记",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生端"
                ],
                "summary": "获取可用考试列表",
                "responses": {
                    "200": {
                        "description": "考试列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.StudentExamItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/student/records": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "查询指定考生的所有考试成绩记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生端"
                ],
                "summary": "获取考生历史成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考生姓名",
                        "name": "studentName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成绩列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.StudentRecordItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ExamQuestionsData": {
            "type": "object",
            "properties": {
                "allowRepeat": {
                    "type": "boolean"
                },
                "canViewAnswer": {
                    "type": "boolean"
                },
                "correctCount": {
                    "type": "integer"
                },
                "examId": {
                    "type": "integer"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExamQuestionItem"
                    }
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResultItem"
                    }
                },
                "title": {
                    "type": "string"
                },
                "totalQuestions": {
                    "type": "integer"
                },
                "totalScore": {
                    "description": "已完成时包含的字段",
                    "type": "number"
                },
                "wrongCount": {
                    "type": "integer"
                }
            }
        },
...
        "models.PagedData": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "输入 \"Bearer {token}\" 或通过 Cookie exam_token 传递",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}